+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.2 (UNRELEASED, 2018)
 
 * Add Travis configuration, `make lint` and git precommit hook
+* Fix `--help` displaying "pflag: help requested"
 * Fix issue with make not recompiling when source files changed
 * Fix issue with `make test` always returning true even when tests fail
 * Fix race condition that could cause failures due to astro downloading the